saas_onb_tooling/requirements.txt
requests 
regex 
ldap3 
apache-airflow
GitPython 
python-gitlab